requirements.txt
python-dotenv
httpx
requests
openai
pvporcupine
pyaudio
speechrecognition
edge-tts
pygame
rich
playwright
pycaw
comtypes
screen_brightness_control
keyboard
pyautogui
pillow
chromadb
sentence-transformers
pypdf
apscheduler
dateparser
watchdog
